pyproject.toml
[project]
name = "media-resolver-mcp"
version = "0.1.0"
description = "MCP server for resolving and playing media via Mopidy and Home Assistant"
authors = [{name = "Media Resolver Team"}]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # MCP Server Framework
    "fastmcp>=2.0.0",

    # Web Framework for Admin UI
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "jinja2>=3.1.4",

    # LangChain for LLM integration
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",

    # HTTP clients
    "httpx>=0.27.0",
    "aiohttp>=3.10.0",

    # Configuration and data handling
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "pyyaml>=6.0.2",

    # RSS parsing for podcasts
    "feedparser>=6.0.11",
    "python-dateutil>=2.9.0",

    # Utilities
    "python-dotenv>=1.0.1",
    "structlog>=24.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.27.0",  # For testing
    "black>=24.10.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]

[project.scripts]
media-resolver = "media_resolver.server:main"

[build-system]
requires = ["setuptools>=75.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "pydantic.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
